+++ bindings/pylucid/gui.py
@@ -12,9 +12,10 @@ import numpy as np
 from cachelib import FileSystemCache
 from flask import Blueprint, Flask, Response, request, send_from_directory, session
 from flask_cors import CORS
-from flask_session import Session
 from jsonschema import ValidationError
 from pyparsing import ParseException
+
+from flask_session import Session
 
 from . import CAPABILITIES
 from .__main__ import scenario_config